Added welcome page
A welcome page was added with an about section where the user can read
about what Las Nubes is, also added a Contact section (there is no info
in neither of this sections I'll add it later). The page has a sign in
and a sign up buttons, so when the user clicks any of this buttons the
respective form will revealed to the user (ether to sign up or sign in).