Introducing FilesHandler
File handler is a class that will be used to upload, download and get
the file list from all the cloud providers. This FilesHandler class will
now to what cloud provider upload a given file, from where download it
and how to list all cloud providers files, all this based on the user
configuration (not defined yet). Also the application can now upload
files to dropbox.