Created status skeleton and solved bug with session creation
When user starts an upload a status pane will show to report the upload
progress(at this point it just show up it do not report progress). Also
solved a bug with session creation, it was throwing a no method error
I delete the lines that were causing that problem because those lines
are not needed anymore.